requirements.txt
pyTelegramBotAPI>=4.0.0
python-whois>=0.8.0
python-dotenv